+++ website/src/examples/basic_navbar.rs
@@ -8,10 +8,14 @@ use leptos_bulma::components::{
 pub fn BasicNavbar() -> impl IntoView {
     let burger_is_active = create_rw_signal(false);
 
+    let show_alert = move |_| {
+        let _ = window().alert_with_message("Item clicked");
+    };
+
     view! {
         <BNavbar class="is-primary has-shadow block">
             <BNavbarBrand>
-                <BNavbarItem class="is-size-4" href="/">
+                <BNavbarItem class="is-size-4" on:click=show_alert>
                     "Leptos Bulma"
                 </BNavbarItem>
 
@@ -20,19 +24,19 @@ pub fn BasicNavbar() -> impl IntoView {
 
             <BNavbarMenu is_active=burger_is_active>
                 <BNavbarStart>
-                    <BNavbarItem>"Item"</BNavbarItem>
+                    <BNavbarItem on:click=show_alert>"Item"</BNavbarItem>
                 </BNavbarStart>
 
                 <BNavbarEnd>
-                    <BNavbarItem>"Item"</BNavbarItem>
+                    <BNavbarItem on:click=show_alert>"Item"</BNavbarItem>
 
                     <BNavbarItemDropdown
                         dropdown_class="is-right"
                         trigger=|| view! { <span class="has-text-weight-bold">"Dropdown item"</span> }
                     >
-                        <BNavbarItem>"Item"</BNavbarItem>
+                        <BNavbarItem on:click=show_alert>"Item"</BNavbarItem>
                         <BNavbarDivider/>
-                        <BNavbarItem>"Item"</BNavbarItem>
+                        <BNavbarItem on:click=show_alert>"Item"</BNavbarItem>
                     </BNavbarItemDropdown>
                 </BNavbarEnd>
             </BNavbarMenu>